+++ src/uci_functionality/move.cpp
@@ -8,7 +8,7 @@ void UCI::inputMove(string input) {
 	std::vector<std::string> inVec;
 	UCI::splitString(input, &inVec);
 
-	short pos(1);
+	size_t pos(1);
 	while (pos != inVec.size()) {
 		cs.move(cs.notationToMove(inVec[pos]));
 		pos += 1;
