+++ src/uci_functionality/position.cpp
@@ -11,7 +11,7 @@ void UCI::inputPosition(string input) {
 	std::vector<std::string> inVec;
 	UCI::splitString(input, &inVec);
 
-	short pos(2);
+	size_t pos(2);
 
 	// Set position
 	if (inVec[1] == "fen") {
